// packages/request-node/src/types.ts
export interface Adapter {
  (request: HttpRequest): Promise<HttpResponse>;
}
// ...
export interface HttpRequest {
  /**
   * The URL of a resource to request.
   */
  readonly url: string;
  /**
   * The HTTP method to use for a request.
   */
  readonly method: string;
  /**
   * The request headers.
   */
  readonly headers?: HttpHeaders | null;
  /**
   * The request body.
   */
  readonly body?: BodyDataType | null;
  /**
   * The event emitter to be notified of request or response events.
   */
  readonly eventEmitter?: EventEmitter | null;
  /**
   * Any additional request options.
   */
  readonly options?: HttpRequestOptions | null;
}

export interface HttpRequestOptions extends SecureContextOptions {
  readonly timeout?: number;
  readonly agent?: AnyAgent | ((url: string) => AnyAgent);
  readonly rejectUnauthorized?: boolean;
}

export type BodyDataType = string | Buffer | Readable | Streamable;

/**
 * Represents response of a web request, if completed and parsed successfully.
 */
export interface HttpResponse {
  /**
   * Contains value indicating whether the response was successful (status in
   * the range 200-299) or not.
   */
  readonly ok: boolean;
  /**
   * Response status code.
   */
  readonly status: number;
  /**
   * Response status text.
   */
  readonly statusText: string;
  /**
   * Response URL after all redirects, if any, or the original URL.
   */
  readonly url: string;
  /**
   * The bag of all response headers.
   */
  readonly headers: HttpHeaders;
  /**
   * Response body.
   */
  readonly body: Body;

  /**
   * Discards response body, closes the incoming body stream.
   *
   * This method can be called, for example, when the received response is not
   * of the expected MIME type. In such a situation the client may discards the
   * body and throw an error.
   */
  abort(): void;
}

export interface UploadProgressEvent {
  readonly type: "upload";
  readonly loaded: number;
  readonly total: number | null;
}

export interface DownloadProgressEvent {
  readonly type: "download";
  readonly loaded: number;
  readonly total: number | null;
}

export type NameValueEntries = readonly (readonly [string, unknown])[];

export type AnyAgent = HttpAgent | HttpsAgent | boolean;